adding complete functionality

+++ apps/core/views.py
@@ -2,15 +2,16 @@ from multiprocessing import context
 from django.forms import ValidationError
 import requests
 import pygal
-from django.http import HttpResponseRedirect
+from django.contrib import messages
+from django.http import Http404, HttpResponseRedirect
 from django.shortcuts import render, redirect
 from django.core.paginator import Paginator
 from django.contrib.auth.decorators import login_required
 from django.core.exceptions import SuspiciousOperation
 
 from apps.accounts.models import User
-from apps.core.models import Book, ReadingList, Chart, StatesList
-from apps.core.forms import AddBookForm, AddReadingListForm, AddChartForm, EditChartForm, AddStateForm
+from apps.core.models import ChartProperties, Chart, StatesList, Voted, state_abbr_list
+from apps.core.forms import AddChartForm, EditChartForm, AddStateForm
 
 
 # Start of my changes
@@ -34,33 +35,37 @@ def chart_function(chart):
     print('Using Editing Function')
 
     states_for_chart = StatesList.objects.filter(chart_id=chart.id)
+    properties_for_chart = ChartProperties.objects.get(chart_id=chart.id)
+    
     for state in states_for_chart:
-        print(' Getting data frrrom chart',state.chart_id)
+        # print(' Getting data frrrom chart',state.chart_id)
         print('The state is: ', state.state)
     state_dict = {}
-    if chart.chart_type == 'bar':
-        plot = pygal.Bar()
+    title = properties_for_chart.title
+    if properties_for_chart.chart_type == 'bar':
+        plot = pygal.Bar(title=title)
     else:
-        plot = pygal.Pie()
+        plot = pygal.Pie(title=title)
 
     for state in states_for_chart:
-        print('------------------')
+        # print('------------------')
         #ensuring valid url
-        print('url', 'https://api.covidtracking.com/v1/states/'+ state.state.lower() +'/'+ chart.year + chart.month + str(chart.day) +'.json')
-        print('------------------')
-        response = requests.get('https://api.covidtracking.com/v1/states/'+ state.state.lower() +'/'+ chart.year + chart.month + str(chart.day) +'.json')
+        # print('url', 'https://api.covidtracking.com/v1/states/'+ state.state.lower() +'/'+ properties_for_chart.year + properties_for_chart.month + str(properties_for_chart.day) +'.json')
+        # print('------------------')
+        response = requests.get('https://api.covidtracking.com/v1/states/'+ state.state.lower() +'/'+ properties_for_chart.year + properties_for_chart.month + str(properties_for_chart.day) +'.json')
         state_data = response.json()
-        state_dict[state_data['state']] = state_data[chart.filter_field]
-        print(state_dict)
+        state_dict[state_data['state']] = state_data[properties_for_chart.filter_field]
+        # print(state_dict)
     
     sorted_state_keys = sorted(state_dict, key=state_dict.get, reverse=True)
-    print(sorted_state_keys)
+    # print(sorted_state_keys)
 
     for k in sorted_state_keys:
         v =state_dict[k]
         value = int(v)
         label = str(k)
         plot.add(label, value)
+    
     plots_svg = plot.render(is_unicode=True)
     chart.plot_entry = plots_svg
     chart.save()
@@ -68,14 +73,11 @@ def chart_function(chart):
          
 
 def user_page(request, username):
-
-    user = User.objects.get(username=username)
-    
+    user = User.objects.get(username=username)  
 # Checkign to see if the user has created any charts
 # If not they are redirected to the create chart page
 # If they do have at least one chart they will stay on their page and their charts will be displayed
-# Maybe going to have the svg be saved in the model instead of being regenerated upon refresh
-   #Checks how many charts the user has plotted
+# Checks how many charts the user has plotted
     if  len(Chart.objects.filter(creator_user_id = user.id)) != 0:
         charts = Chart.objects.filter(creator_user_id = user.id)
         print( len(Chart.objects.filter(creator_user_id = user.id)))
@@ -107,24 +109,32 @@ def user_page(request, username):
 
 @login_required
 def create_chart(request, username):
-    
+    user = User.objects.get(username=username)
+
     if request.method == 'POST':
         # Create a form instance and populate it with data from the request
         form = AddChartForm(request.POST)
         if form.is_valid():
             # If we had omitted commit=False, then the user would not have been
             # properly set-up
-            new_chart = form.save(commit=False)
+            # Creating chart instance in the DB so properties and stateslist can be based off it
+            new_chart = Chart.objects.create(creator_user_id = user.id)
             new_chart.creator_user = request.user
+
+            new_chart_prop = form.save(commit=False)
+            new_chart_prop.chart_id = new_chart.id
             print('---------------')
-            print('The creator user: ', new_chart.creator_user)
+            print('The creator user: ', new_chart_prop)
             print('---------------')
-            new_chart.save()
+            new_chart_prop.save()
             # Creating the linked entry in the statesList table
-            stateList = StatesList.objects.create(chart = new_chart, state = new_chart.stateAbr)
+            stateList = StatesList.objects.create(chart = new_chart, state = new_chart_prop.stateAbr)
 
             chart = Chart.objects.get(id = new_chart.id)
+            chart.title = new_chart_prop.title
+            chart.save()
             plots_svg = chart_function(chart)
+  
             return redirect('/charts/'+username+'/')
     else:
         # if a GET  we'll create a blank form
@@ -140,47 +150,66 @@ def edit_chart(request, username, chart_id):
     print('EDITING CHART VIEW')
     #Getting chart that the user selected
     chart = Chart.objects.get(id = chart_id)
+    chart_prop = ChartProperties.objects.get(chart_id = chart_id)
     #Get the stateslist instance linked to the chart id so the user can enter data
    
     url = '/charts/'+str(username)+'/'+str(chart_id)+'/'
     print('------------------------')
     print('The url is :', url)
     print('the chart is: ', chart.id)
+    print('the title is: ',chart_prop.chart_id)
 
     print('------------------------')
 
     if request.method == 'POST':
         # Create a form instance for the non-state fields
         #  and populate it with data from the request
-        formChart = EditChartForm(request.POST, instance=chart)
-
+        formChartProp = EditChartForm(request.POST, instance=chart_prop)
+        
         #Create form instance for adding state to plot
         formStateList = AddStateForm(request.POST)
+
         if formStateList.is_valid():
+            print('STATE FORM VALID')
             #Updates values in the stateList database
             new_state = formStateList.save(commit=False)
+            print(new_state)
             #Check if already in chart
             existing_states = StatesList.objects.filter(chart_id = chart.id)
+            if new_state.state !='':
+                for existing_state in existing_states:
+                    print(existing_state.state.lower())
+                    #Delete New Entry as to not have duplicates 
+                    if new_state.state.lower() not in state_abbr_list:
+                        messages.error(request,'Please enter a valid state')
+                        return HttpResponseRedirect(request.META.get('HTTP_REFERER'))
+                    if existing_state.state.lower()!= new_state.state.lower():
+                        new_state.chart = chart
+                        new_state.save()
+                    else:
+                        new_state.delete() 
+        else:
+            print('not valid', formStateList.errors)
 
-            for existing_state in existing_states:
-                print(existing_state.state.lower())
-                if existing_state.state.lower() != new_state.state.lower():
-                    new_state.chart = chart
-                    new_state.save()
-                #Delete New Entry as to not have duplicates  
-                else:
-                    new_state.delete()
-            
-
-        if formChart.is_valid():
-            chart = formChart.save(commit=False)
+        print('-----------------') 
+        print('break between forms')
+        print('-----------------')  
+        if formChartProp.is_valid():
+            new_chart_prop = formChartProp.save(commit=False)
+            print('CHART FORM INFO IS VALID')
+            print('the title is: ', new_chart_prop.title)
+            new_chart_prop.save()
+            chart.title = new_chart_prop.title
             chart.save()
             return redirect(url)
-        
+        else:
+            print('not valid', formChartProp.is_valid())
+            print(formChartProp.errors)    
 
+        
     else:
         # A GET, create a pre-filled form with the instance.
-        formChart = EditChartForm(instance=chart)
+        formChartProp = EditChartForm(instance=chart_prop)
         formStateList = AddStateForm()
 
     plots_svg = chart_function(chart)
@@ -195,7 +224,7 @@ def edit_chart(request, username, chart_id):
         })
 
     context = {
-        'formChart' : formChart,
+        'formChartProp' : formChartProp,
         'formState' : formStateList,
         'rendered_chart': chart.plot_entry,
         'states': states,
@@ -222,28 +251,26 @@ def delete_chart(request, username, chart_id):
     if chart.creator_user_id != request.user.id:
        raise SuspiciousOperation('Attempted to delete wrong list')
 
-    return redirect('/charts/'+username+'/')
+    return HttpResponseRedirect(request.META.get('HTTP_REFERER'))
+    #redirect('/charts/'+username+'/')
 # Existing content/functionality to be removed/phased out
-def reading_list_home(request):
-    # R in CRUD --- READ ReadingLists from database
+def chart_home(request):
     plots_svg = []
     charts = Chart.objects.all()
    
     for chart in charts:
         plot_svg = chart.plot_entry
         plots_svg.append(plot_svg)
-        
-    
-    #.select_related('creator_user')
 
-    # Let's sort by their "score"
-    #reading_lists = reading_lists.order_by('-score')
+    #sort by votes
+    charts = charts.order_by('-votes')
 
     # And "paginate" the results (split them into pages)
-    # https://docs.djangoproject.com/en/3.0/topics/pagination/
+    
     page_number = request.GET.get('page', 1)
     paginator = Paginator(charts, 4)
     results_page = paginator.page(page_number)
+    print(results_page)
 
     context = {
         'results_page': results_page,
@@ -251,98 +278,38 @@ def reading_list_home(request):
     }
     return render(request, 'pages/home.html', context)
 
-def reading_list_details(request, list_id):
-    # R in CRUD --- READ a single ReadingList & its books from database
-    reading_list_requested = ReadingList.objects.get(id=list_id)
-
-    # Count views of pages for determining what's popular
-    reading_list_requested.increment_views()
-
-    books = Book.objects.filter(reading_list=reading_list_requested)
-    context = {
-        'reading_list': reading_list_requested,
-        'all_books': books,
-    }
-    return render(request, 'pages/details.html', context)
-
 @login_required
-def reading_list_create(request):
-    if request.method == 'POST':
-        # Create a form instance and populate it with data from the request
-        form = AddReadingListForm(request.POST)
-        if form.is_valid():
-            # If we had omitted commit=False, then the user would not have been
-            # properly set-up
-            new_reading_list = form.save(commit=False)
-            new_reading_list.creator_user = request.user
-            new_reading_list.save()
-            return redirect(new_reading_list.get_absolute_url())
-    else:
-        # if a GET  we'll create a blank form
-        form = AddReadingListForm()
-    context = {
-        'form': form,
-        'form_title': 'New book list',
-    }
-    return render(request, 'pages/form_page.html', context)
+def chart_vote_up(request, user_id, chart_id):
 
+    user = User.objects.get(id=user_id)
+    c1 = Chart.objects.get(id=chart_id)
+    v1 = Voted(has_voted=False)
 
-@login_required
-def reading_list_delete(request, list_id):
-    # D in CRUD --- DELETE reading list from database
-    readinglist = ReadingList.objects.get(id=list_id)
-
-    # Prevent users who are not the owner user from deleting this
-    #if readinglist.creator_user != request.user:
-    #    raise SuspiciousOperation('Attempted to delete wrong list')
-
-    readinglist.delete()
+    #check if user has voted, and if its for an upvote
+    # if Voted.objects.get(user=user_id, chart=chart_id) == 0:
+    #     vote = Voted(has_voted=True)
+    #     vote.vote = 'up'
+    #     vote.chart.add(chart_id)
+    #     vote.user.add(user_id)
+    #     vote.save()
+    # else:
+    #     vote = Voted.objects.get(user=user_id, chart=chart_id)
+    #     vote.vote = 'up'
+    #     vote.save()
+    
+    chart = Chart.objects.get(id=chart_id)
+    chart.votes = chart.votes + 1
+    chart.save()
     return redirect('/')
 
-
 @login_required
-def reading_list_create_book(request, list_id):
-    # C in CRUD --- CREATE books in database
-    reading_list_requested = ReadingList.objects.get(id=list_id)
+def chart_vote_down(request, user_id, chart_id):
+    # U in CRUD, update the reading list based on
+    chart = Chart.objects.get(id=chart_id)
+    chart.votes = chart.votes - 1
+    chart.save()
+    return redirect('/')
 
-    # TODO: BONUS CHALLENGE - Uncomment this to fix the security defect
-    #if reading_list_requested.creator_user != request.user:
-    #    raise SuspiciousOperation('Attempted to add book to wrong list')
-
-    if request.method == 'POST':
-        # Create a form instance and populate it with data from the request
-        form = AddBookForm(request.POST)
-        if form.is_valid():
-            book = form.save(commit=False)
-            book.reading_list = reading_list_requested
-            book.save()
-
-            # Redirect back to the reading list we were at
-            return redirect(reading_list_requested.get_absolute_url())
-    else:
-        # if a GET  we'll create a blank form
-        form = AddBookForm()
-    context = {
-        'form': form,
-        'form_title': 'Add book',
-    }
-    return render(request, 'pages/form_page.html', context)
-
-
-@login_required
-def reading_list_delete_book(request, book_id):
-    # D in CRUD, delete book
-    book = Book.objects.get(id=book_id)
-
-    # Ensure that the creator of the reading list of the book is indeed the
-    # person requesting that the book be deleted
-    # TODO: Challenge 4 Uncomment this to fix the security defect
-    #if book.reading_list.creator_user != request.user:
-    #    raise SuspiciousOperation('Attempted to delete book to wrong list')
-
-    book.delete()
-
-    return redirect(book.reading_list.get_absolute_url())
